+++ src/Aks/Aks.Test/ScenarioTests/KubernetesTests.cs
@@ -198,5 +198,12 @@ namespace Commands.Aks.Test.ScenarioTests
         {
             TestRunner.RunTestScript("Test-PodSubnetID");
         }
+
+        [Fact]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        public void TestEnableOidcIssuer()
+        {
+            TestRunner.RunTestScript("Test-EnableOidcIssuer");
+        }
     }
 }